requirements.txt
Flask==2.3.4
torch==2.2.0
torchvision==0.18.0
Pillow==10.1.0
werkzeug==2.3.8
